refactor(statement): remove bootstrap directory
