+++ apps/statement/src/21-mobx-basic/index.tsx
@@ -1,6 +1,6 @@
 import { observer } from 'mobx-react';
 import { BrowserRouter } from 'react-router-dom';
-import { App } from './bootstraps/App';
+import { App } from './App';
 
 export const MobxBasicApp = observer(() => (
   <BrowserRouter>
