+++ apps/statement/src/22-mobx-hooks/index.tsx
@@ -1,6 +1,6 @@
 import { configure } from 'mobx';
 import { BrowserRouter } from 'react-router-dom';
-import { App } from './bootstraps/App';
+import { App } from './App';
 
 export const MobxHooksApp = () => (
   <BrowserRouter>
