+++ apps/statement/src/31-unstated-basic/App.tsx
@@ -4,19 +4,19 @@ import { PageTransition } from '@learn-react/core/src/components/utils/PageTrans
 import { gutter } from '@learn-react/core/src/helpers/Style';
 import { type ComponentProps } from 'react';
 import { Route } from 'react-router-dom';
-import { Router } from '../../@core/constants/Router';
-import { HomePage } from '../pages/Home';
-import { ListPage } from '../pages/ListPage';
-import { NotFoundPage } from '../pages/NotFoundPage';
-import { Profile } from '../pages/profiles';
+import { Router } from '../@core/constants/Router';
+import { HomePage } from './pages/HomePage';
+import { ListPage } from './pages/ListPage';
+import { NotFoundPage } from './pages/NotFoundPage';
+import { ProfilePage } from './pages/profiles';
 
 export const App = () => (
-  <div className={baseStyle}>
-    <Sidebar title="MobX Basic" items={linkItems} />
-    <div className={contentStyle}>
+  <div className={styleBase}>
+    <Sidebar title="Unstated | Statement" items={linkItems} />
+    <div className={styleContent}>
       <PageTransition>
         <Route path={Router.paths.home} element={<HomePage />} />
-        <Route path={`${Router.paths.profile}/*`} element={<Profile />} />
+        <Route path={`${Router.paths.profile}/*`} element={<ProfilePage />} />
         <Route path={Router.paths.list} element={<ListPage />} />
         <Route element={<NotFoundPage />} />
       </PageTransition>
@@ -25,10 +25,6 @@ export const App = () => (
 );
 
 const linkItems: ComponentProps<typeof Sidebar>['items'] = [
-  {
-    label: 'Home',
-    to: Router.paths.home,
-  },
   {
     label: 'Profile',
     items: [
@@ -48,12 +44,12 @@ const linkItems: ComponentProps<typeof Sidebar>['items'] = [
   },
 ];
 
-const baseStyle = css`
+const styleBase = css`
   display: flex;
   width: 100%;
 `;
 
-const contentStyle = css`
+const styleContent = css`
   flex-grow: 1;
   height: 100dvh;
   padding: ${gutter(4)};
